requirements.txt
tensorflow
numpy
pandas
scikit-learn
keras
joblib
sentence-transformers
faiss-cpu
tqdm